requirements.txt
ws4py
gevent
pyserial
